backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels[daphne]
spotipy
requests
PyJWT
python-dotenv
pillow
